% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helliniko}
\alias{helliniko}
\title{Helliniko maximum intensities}
\format{A tibble with 228 rows and 2 variables:
\describe{
  \item{duration}{The duration of the maximum intensity for a specific year}
  \item{intensity}{Maximum intensity for a specific year and duration, in
  mm/hr}
}}
\source{
D. Koutsoyiannis, Statistical Hydrology, Edition 4, 312 pages,
doi:10.13140/RG.2.1.5118.2325, National Technical University of Athens,
Athens, 1997.
}
\usage{
helliniko
}
\description{
A 30 year (1957-58 to 1986-87) data record of the Helliniko recording
station (located at the Helliniko airport, Athens). The durations range from
5/60 h  (5 min) to 24 h. Owing to missing values, the sample size for some durations is lower than 30.
}
\keyword{datasets}
