% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_aggregate.R
\name{hyet_aggregate}
\alias{hyet_aggregate}
\title{Aggregate a hyetograph using a time step}
\usage{
hyet_aggregate(hyet, time_step, ts_unit)
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function.}

\item{time_step}{an integer specifying the time step value of the aggregated
hyetograph.}

\item{ts_unit}{a character string specifying the time unit of the aggregated
hyetograph. Valid units are "mins", "hours", "days", "months", "quarter" or
"year".}
}
\value{
a tibble with the aggregated hyetograph.
}
\description{
\code{hyet_aggregate} uses a predefined time step value to
aggregate precipitation records in an hyetograph.
}
\examples{

prec30min <- hyet_aggregate(prec5min, 30, "mins")

}
