% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_arrive.R
\name{hyet_arrive}
\alias{hyet_arrive}
\title{Compute interarrival and dry times duration from independent rainstorms}
\usage{
hyet_arrive(storms)
}
\arguments{
\item{storms}{a grouped dataframe with independent rainstorms.}
}
\value{
a tibble with the duration, total precipitation height
and interarrival and dry times in hours of the independent rainstorms.
}
\description{
\code{hyet_interar} uses independent rainstorms that comes
from the \code{hyet_split} function and computes their interarrival and
dry time duration.
Returns an error if \code{storms} is not a valid grouped hyetograph.
}
\note{
Please note that the interarrival and dry time durations for the first
rainstorms is set to zero.
}
\examples{
# create time series with 1 hours time step
time_step <- 1
ts_unit <- "hours"
len <- 30
crit_dur <- rep(6, 12)

hyet <- tibble::tibble(
  date = seq(
    from = as.POSIXct(0, origin = "2018-01-01"), length.out = len,
    by = paste(time_step, ts_unit)
  ),
  prec = runif(len, 1, 5)
)

# create 2 dry durations of 6 hours
hyet$prec[10:15] <- 0
hyet$prec[21:26] <- 0

# add some zeros
hyet$prec[2:5] <- 0

# split hyet and compute interarrival times
hyet \%>\%
  hyet_split(time_step, ts_unit, crit_dur) \%>\%
  hyet_arrive()
}
