% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_fill.R
\name{hyet_fill}
\alias{hyet_fill}
\title{Fill a hyetograph with missing date values}
\usage{
hyet_fill(hyet, time_step = 5, ts_unit = "mins")
}
\arguments{
\item{hyet}{a hyetograph from \code{hyet_create} function}

\item{time_step}{an integer that represents hyetograph's time-step.}

\item{ts_unit}{a character string specifying a time unit. Valid units are
"mins" and "hours".}
}
\value{
a tibble with the variables \code{date} and \code{prec}
}
\description{
\code{hyet_fill} fills missing date values in a hyetograph
with NA values.  Returns an error if \code{hyet} is not a valid hyetograph or
\code{time_step} is a not a numeric value.
}
\examples{

# create date and precipitation values using 5 minutes time-step
prec_date <- seq(from = as.POSIXct(0, origin = "2018-01-01"),
                  length.out =  100,
                  by = "5 mins")
set.seed(1)
prec_values <-round(runif(100,0,10),1)

# create hyetograph
hyet <- hyet_create(prec_date, prec_values)

# remove some random values from hyetograph
hyet_miss <- hyet[-sample(100,30), ]

# fill hyetograph
hyet_fill(hyet_miss, 5, "mins")
}
