% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_split.R
\name{hyet_split}
\alias{hyet_split}
\title{Split an hyetograph to independent rainstorms}
\usage{
hyet_split(hyet, time_step, ts_unit, crit_dur = rep(6, 12))
}
\arguments{
\item{hyet}{an hyetograph from \code{hyet_create} function.}

\item{time_step}{a numeric value that represents the time-step.}

\item{ts_unit}{a character string specifying the time unit. Valid values
are "mins" and "hours".}

\item{crit_dur}{a vector of 12 numeric values in hours. This vector
represents the monthly critical time duration of no precipitation that is
used to separate the hyetograph to independent rainstorms. Default value for
all months is six hours.}
}
\value{
a grouped dataframe with independent rainstorms.
}
\description{
\code{hyet_split} uses a predefined vector with monthly values
of fixed critical time duration of no precipitation to split an hyetograph
to independent rainstorms.
Returns an error if \code{hyet} is not a valid hyetograph.
}
\examples{

# create time series with 1 hours time step
time_step <- 30
ts_unit <- "mins"
len <- 90
crit_dur = rep(6,12)
hyet <- tibble::tibble(
  date = seq(
    from = as.POSIXct(0, origin = "2018-01-01"), length.out = len,
    by = paste(time_step, ts_unit)
  ),
  prec = runif(len, 1, 5)
)
hyet$prec[10:23] <- 0
hyet$prec[41:54] <- 0

# split hyetograph to storms
storms <- hyet \%>\%
  hyet_split(time_step, ts_unit, crit_dur)

}
\references{
Restrepo-Posada, Pedro J., and Peter S. Eagleson. "Identification of
independent rainstorms." Journal of Hydrology 55.1-4 (1982): 303-319.
}
