% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyetor-package.R
\docType{package}
\name{hyetor}
\alias{hyetor}
\alias{_PACKAGE}
\alias{hyetor-package}
\title{hyetor: Analyzes fixed interval precipitation records.}
\description{
This package provides a collection of tools that can be used to analyze fixed
 interval precipitation time series:
 \itemize{
   \item {Aggregate precipitation time series values.}
   \item {Split precipitation records to independent rainstorms.}
   \item {Compile Unitless Cumulative Hyetographs.}
   \item {Find maximum rainfall intensities for various time steps.}
   \item {Calculate rainfall erosivity values.}
   \item {Create missing values summaries.}
   }
}
\section{Data Sources}{


The data are retrieved from:
 \url{http://www.hydroscope.gr/}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kvantas/hyetor}
  \item \url{https://kvantas.github.io/hyetor/}
  \item Report bugs at \url{https://github.com/kvantas/hyetor/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (0000-0001-6387-8791)

}
