% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prec5min}
\alias{prec5min}
\title{Five minutes precipitation time series}
\format{A tibble with 48,209 rows and 2 variables:
\describe{
  \item{date}{The time series' dates}
  \item{prec}{Precipitation, in mm}
}}
\source{
http://kyy.hydroscope.gr/timeseries/d/1432/
}
\usage{
prec5min
}
\description{
This time series comes from the weather station 'Arna' in Greece with
coordinates 36.88002, 22.41292 (ETRS89) and altitude 779 m.
The owner of that weather station is the Ministry of Environment and Energy.
Precipitation's units of measurement is in mm and the time step is five
minutes.
}
\keyword{datasets}
