% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyet_split.R
\name{hyet_split}
\alias{hyet_split}
\title{Split an hyetograph to rainstorms using a fixed critical time duration of
no precipitation}
\usage{
hyet_split(hyet, time_step = 5, ts_unit = "mins", crit_dur = rep(6,
  12))
}
\arguments{
\item{hyet}{hyet an hyetograph from \code{hyet_create} function.}

\item{time_step}{an integer that represents hyetograph's time-step.}

\item{ts_unit}{a character string specifying a time unit. Valid units are
"mins" and "hours".}

\item{crit_dur}{an integer or a vector of 12 integers that represent the
critical time duration of no precipitation that is used to separate the
hyetograph to independent rainstorms.}
}
\value{
a grouped dataframe
}
\description{
Split an hyetograph to rainstorms using a fixed critical time duration of
no precipitation
}
\examples{

# create time series with 1 hours time step
time_step <- 30
ts_unit <- "mins"
len <- 90
crit_dur = rep(6,12)
hyet <- tibble::tibble(
  date = seq(
    from = as.POSIXct(0, origin = "2018-01-01"), length.out = len,
    by = paste(time_step, ts_unit)
  ),
  prec = runif(len, 1, 5)
)
hyet$prec[10:23] <- 0
hyet$prec[41:54] <- 0

# split hyet
storms <- hyet_split(hyet, time_step, ts_unit, crit_dur)

}
